Added Pause Menu to the player prefab which has 3 buttons: *Resume *Options (non-functional) *Quit (returns to main menu Scene)
I modified the mouselook and character scripts to disable movement
and looking while paused (however the game state continues running)

To use this new pause menu in a scene, you need to revert the player
to the Prefab defaults, then add a SceneLoader Object to your Scene.
Connect the SceneLoader GameObject to the variable in the player's
"Pause Menu" Sub-Object.

FAILURE TO LINK SCENELOADER RESULTS IN PLAYER FALLING THROUGH FLOOR